requirements.txt
flask==2.0.1
pyTelegramBotAPI==4.0.0
pandas==1.3.3
matplotlib==3.4.3
seaborn==0.11.2
numpy==1.21.2
scikit-learn==1.0
requests==2.26.0
python-dotenv==0.19.1